requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.9
marshmallow==3.5.1
paho-mqtt==1.5.0
requests==2.23.0
urllib3==1.25.8
xmltodict==0.13.0
dataclasses==0.8; python_version < "3.7"
